SRMD_nlp/prevent_dtree/requirement.txt
python>=3.6
scikit-learn>=1.0.2
pandas>=1.3.5
numpy>=1.21.6
matplotlib>=3.5.3
seaborn>=0.11.2
pydotplus>=2.0.2
joblib>=1.1.0
imblearn>=0.0
graphviz>=0.20.1
PyYAML>=6.0
scipy>=1.7.3